feature: added 'comments' functionality

+++ main-service/src/main/java/ru/practicum/main_service/service/compilation/CompilationServiceImpl.java
@@ -3,12 +3,13 @@ package ru.practicum.main_service.service.compilation;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.main_service.dto.compilation.CompilationDto;
 import ru.practicum.main_service.dto.compilation.NewCompilationDto;
 import ru.practicum.main_service.dto.compilation.UpdateCompilationRequest;
 import ru.practicum.main_service.entity.Compilation;
 import ru.practicum.main_service.entity.Event;
-import ru.practicum.main_service.exceptions.CompilationNotExistException;
+import ru.practicum.main_service.exceptions.NotFoundException;
 import ru.practicum.main_service.mapper.CompilationMapper;
 import ru.practicum.main_service.repository.CompilationRepository;
 import ru.practicum.main_service.repository.EventRepository;
@@ -19,7 +20,6 @@ import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Predicate;
 import javax.persistence.criteria.Root;
-import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
@@ -28,6 +28,7 @@ import java.util.Set;
 @Service
 @RequiredArgsConstructor
 @Slf4j
+@Transactional(readOnly = true)
 public class CompilationServiceImpl implements CompilationService {
     private final EventRepository eventRepository;
     private final EventService eventService;
@@ -48,14 +49,18 @@ public class CompilationServiceImpl implements CompilationService {
         compilation.setTitle(newCompilationDto.getTitle());
 
         Compilation savedCompilation = compilationRepository.save(compilation);
-        log.debug("Compilation is created");
+        log.info("Compilation with title {} is created", newCompilationDto.getTitle());
         setView(savedCompilation);
         return mapper.mapToCompilationDto(savedCompilation);
     }
 
     public CompilationDto getCompilation(Long compId) {
         Compilation compilation = compilationRepository.findById(compId)
-                .orElseThrow(() -> new CompilationNotExistException("Compilation doesn't exist"));
+                .orElseThrow(() -> {
+                    log.error("Error fetching compilation with id={}: Compilation doesn't exist", compId);
+                    return new NotFoundException("Compilation doesn't exist");
+                });
+        log.info("Fetched compilation with ID = {}", compId);
         return mapper.mapToCompilationDto(compilation);
     }
 
@@ -83,13 +88,16 @@ public class CompilationServiceImpl implements CompilationService {
                 .setMaxResults(size)
                 .getResultList();
 
+        log.info("Fetched compilations, pinned status: {}", pinned);
+
         return mapper.mapToListCompilationDto(compilations);
     }
 
     @Transactional
     public CompilationDto updateCompilation(Long compId, UpdateCompilationRequest updateCompilationRequest) {
 
-        Compilation oldCompilation = compilationRepository.findById(compId).orElseThrow(() -> new CompilationNotExistException("Can't update compilation - the compilation doesn't exist"));
+        Compilation oldCompilation = compilationRepository.findById(compId).orElseThrow(()
+                -> new NotFoundException("Can't update compilation - the compilation doesn't exist"));
         List<Long> eventsIds = updateCompilationRequest.getEvents();
         if (eventsIds != null) {
             List<Event> events = eventRepository.findAllByIdIn(updateCompilationRequest.getEvents());
@@ -98,19 +106,25 @@ public class CompilationServiceImpl implements CompilationService {
         if (updateCompilationRequest.getPinned() != null) {
             oldCompilation.setPinned(updateCompilationRequest.getPinned());
         }
-        if (updateCompilationRequest.getTitle() != null) {
+        if (updateCompilationRequest.getTitle() != null && !updateCompilationRequest.getTitle().isBlank()) {
             oldCompilation.setTitle(updateCompilationRequest.getTitle());
         }
         Compilation updatedCompilation = compilationRepository.save(oldCompilation);
-        log.debug("Compilation with ID = {} is updated", compId);
+
+        log.info("Compilation with ID = {} is updated", compId);
         setView(updatedCompilation);
+
         return mapper.mapToCompilationDto(updatedCompilation);
     }
 
     @Transactional
     public void deleteCompilation(Long compId) {
+        if (!compilationRepository.existsById(compId)) {
+            log.error("Error deleting compilation with id={}: Compilation doesn't exist", compId);
+            throw new NotFoundException("Compilation doesn't exist");
+        }
         compilationRepository.deleteById(compId);
-        log.debug("Compilation with ID = {} is deleted", compId);
+        log.info("Compilation with ID = {} is deleted", compId);
     }
 
     private void setView(Compilation compilation) {
